Accept FUNCTION SOURCE identifiers mid-line
Reading v8 output from the stdout stream of a child process is sometimes
unreliable, chunks may be missing. The next `FUNCTION SOURCE` can be in
the middle of a previous `FUNCTION SOURCE`, cutting it off. The impact
can be alleviated slightly by accepting `FUNCTION SOURCE` identifiers
that occur in the middle of a line. It's not 100% accurate, because the
missing chunks may contain other `FUNCTION SOURCE` blocks, so we can
still be missing out on stuff. But, at least we don't dump function
source code to the shell anymore.

When this happens, the previous FUNCTION SOURCE block will not have been
closed by an END line, but END lines only set
`stdoutIsPrintOptSourceOutput` to false, and FUNCTION SOURCE lines set
it to true, so we don't have to do anything there.

Related to but does not quite fix #122