Fix init phase detection
The detection will now rely on the `anonymous` function that is used
by node to wrap modules. The very first of such entries is detected
and all other entries must be init frames. This is reliable in a way
that further changes to Node.js internals in the init phase will not
have any impact on the detection.