Fix init detection
This makes sure the init phase is properly detected throughout all
Node.js versions.

A couple of mechanisms work together to achieve that:

1) Get the stack trace while starting Node.js and remove all those
   frames later on.
2) All native modules are detected while initializing and all frames
   before that one including it will also be marked as bootstrapping
   frames.
3) For Node.js 10 a few more stacks are detected by directly checking
   for the `internal/bootstrap.js` file. It would not show up in the
   other two mechanisms.
4) A check for `internalBinding` is in place since this code would
   only execute during native module loading.
5) An explicit check for `NativeModule` is in place since that is the
   name of the internal Node.js module loading to bootstrap core.

This makes sure that pretty much all loading is properly detected.

Besides this, this commit fixes another small bug when using the
`treeDebug` option in combination with the `mapFrames` option. Before
the arguments were not passed through properly.